RabbitMQ communication

+++ rentservice/src/main/java/com/example/rentservice/rentservice/rabbitmq/Receiver.java
@@ -5,6 +5,7 @@ import com.example.rentservice.rentservice.ErrorHandling.ObjectNotFoundException
 import com.example.rentservice.rentservice.Models.RealEstate;
 import com.example.rentservice.rentservice.Repositories.RealEstateRepository;
 import com.example.rentservice.rentservice.Services.RealEstateService;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.amqp.core.AmqpTemplate;
 import org.springframework.amqp.rabbit.annotation.RabbitHandler;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
@@ -20,30 +21,45 @@ public class Receiver {
 
     @Autowired
     private RealEstateService realEstateService;
+    @Autowired
+    private ObjectMapper objectMapper;
 
     @RabbitHandler
     public void receive(String message) throws InvalidRequestException, ObjectNotFoundException {
-        Integer realEstateId = Integer.parseInt(message);
-        System.out.println("Received message = " + realEstateId);
-        var realEstateToDelete = realEstateService.findRealEstateById(realEstateId).getBody();
-        try {
-            realEstateService.findRealEstateById(realEstateId).getBody();
-            realEstateService.deleteRealEstate(realEstateId);
-
-            rabbitTemplate.convertAndSend(RabbitMQConfig.queueName, "Ok " + realEstateId);
-            System.out.println("Sent message with status: Ok " + realEstateId);
+        if(message.contains("customerId")){
+            RealEstate newRealEstate = objectMapper.convertValue(message, RealEstate.class);
+            try{
+                realEstateService.saveRealEstate(newRealEstate);
+                System.out.println("Sent message with status: Ok " + newRealEstate);
+            }
+            catch(Exception ex){
+                rabbitTemplate.convertAndSend(RabbitMQConfig.queueName, "Error " + newRealEstate);
+                System.out.println("Sent message with status: Error " + newRealEstate);
+            }
 
         }
-        catch (ObjectNotFoundException ex) {
-            realEstateService.saveRealEstate(realEstateToDelete);
-            rabbitTemplate.convertAndSend(RabbitMQConfig.queueName, "Error " + realEstateId);
-            System.out.println("Sent message with status: Error " + realEstateId);
-        }
-        catch (Exception ex) {
-            rabbitTemplate.convertAndSend(RabbitMQConfig.queueName, "Error " + realEstateId);
-            System.out.println("Sent message with status: Error " + realEstateId);
-        }
+        else{
+            Integer realEstateId = Integer.parseInt(message);
+            System.out.println("Received message = " + realEstateId);
+            var realEstate = realEstateService.findRealEstateById(realEstateId).getBody();
+            try {
 
+                realEstateService.deleteRealEstate(realEstateId);
+
+                rabbitTemplate.convertAndSend(RabbitMQConfig.queueName, "Ok " + realEstateId);
+                System.out.println("Sent message with status: Ok " + realEstateId);
+
+            }
+            catch (ObjectNotFoundException ex) {
+                realEstateService.saveRealEstate(realEstate);
+                rabbitTemplate.convertAndSend(RabbitMQConfig.queueName, "Error " + realEstateId);
+                System.out.println("Sent message with status: Error " + realEstateId);
+            }
+            catch (Exception ex) {
+                rabbitTemplate.convertAndSend(RabbitMQConfig.queueName, "Error " + realEstateId);
+                System.out.println("Sent message with status: Error " + realEstateId);
+            }
+        }
     }
 
 }
